syncnet-service/requirements.txt
# Core Flask
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# SyncNet dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
tqdm>=4.66.0

# Face detection
# facenet-pytorch>=2.5.3  # Commented - SyncNet uses S3FD detector included in repo

# Utils
python-multipart>=0.0.6
Pillow>=10.0.0

# Video processing (ffmpeg wrapper)
ffmpeg-python>=0.2.0

# --- EfficientNet Dependencies (Added 2025-11-03) ---
efficientnet-pytorch>=0.7.1
timm>=1.0.3
Pillow>=10.0.0
scikit-learn>=1.3.0

# --- Vision Transformer Dependencies (Added 2025-11-03) ---
transformers>=4.30.0

# --- EfficientNetV2-B2 Dependencies (Updated 2025-11-03) ---
# timm>=1.0.3 already included above (supports EfficientNetV2)
